Create app/lsp.rs
